requirements.txt
aniso8601==9.0.1
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
PyJWT==2.6.0
pymongo==4.3.3
pytz==2022.6
six==1.16.0
Werkzeug==2.2.2
zipp==3.11.0
